Only perform arpreply_clear() after consuming an arp packet in handle_arp_response() and when opening a new arp fd.  Do not clear the arpreply buffer and offset on ARP state transitions.
Only perform ARP collision probing when binding a new lease from the
DS_REQUESTING state, or if we have renewed or rebound a lease with a different
IP than we had before.

Resequence the arp_dhcp_packet memcpy() in arp_check() so that the current
IP address is ARP checked rather than the previous one.