Make example s6 script handle empty leasefile properly.
Prior to this change, ndhc would fail to run with a malformed
command error if the leasefile was empty since the -r parameter
requires an argument.

Note that this script requires execline 2.9.1.0 for the
eltest program.  For earlier versions of s6, it should be
fine to use the POSIX test program instead.