arp: Use subnet to determine whether ARP queries can succeed.
Very few modern networks use proxy ARP, so we can assume that
ARP queries will not even be sent by the OS if the destination
isn't on the local subnet.  Thus, we can treat those queries
as having failed by default.

This saves time when fingerprinting the network if the dhcp
agent isn't reachable by ARP.  In the case that proxy arp
actually is present, this just means that we will fail to
fingerprint a server that proxy arp would have allowed us
to fingerprint.