state, arp: Add a fast path for ARP-unreachable gw and agent.
When carrier returns to a downed interface that has a lease,
check to see whether a fingerprint exists, and if it doesn't,
go back to the SELECTING state and reacquire a DHCP lease.

Before, we assumed that there was either an ARP reachable
gateway or DHCP agent, but this isn't necessarily the case,
although the vast majority of networks will have a gateway.

If we have neither, then there's no reliable way to detect
whether we're on the same network when the interface comes
back, so the safe thing is to re-acquire a lease.

In state.c, this just re-uses the ARP send failure path with
a different printed message.

The changes to arp.c are a bit more complex as is always the
case, but are still very localized.