Add a warning about s6 notification
If this option is set to a low file descriptor value such as 3,
then if no notification file descriptor is actually passed to
the ndhc process, then when ndhc tries to perform the requested
s6 notification it will instead close a file descriptor that
may be used for something else.

This problem will manifest as cryptic and hard-to-debug problems.

If a high file descriptor is used, then this problem is less
likely to manifest, but it may cause the kernel to unnecessarily
grow the process file descriptor table larger than it otherwise
should be which may waste some trivial amount of RAM and could
potentially cause some kernel operations that act on the
file descriptor table to be slower than they otherwise would be.

Possibly I could attempt to prevent this configuration issue
by comparing the requested notification file descriptor against
all of the file descriptors that ndhc may poll against and if
it matches one of those file descriptors assuming that the
user has made a configuration error and not proceeding with
the notification.

I am not sure that this feature is actually used by anyone now
that I'm not using it anymore.  If no one speaks up, then I
will likely remove it in a few months or a year.